Moved test files to test/services/ folder
